Moves types to root level and updates imports.
