+++ transaction-processor/src/app/constants.ts
@@ -1,4 +1,4 @@
-import { CardBalance, Transaction } from "./context/types";
+import { CardBalance, Transaction } from "./types/types";
 
 export const headers = ["Account Name", "Card Number", "Transaction Amount", "Transaction Type", "Description", "Target Card Number"];
