+++ transaction-processor/src/app/context/reducer.ts
@@ -1,4 +1,4 @@
-import { Action, TransactionState } from "./types";
+import { Action, TransactionState } from "../types/types";
 
 export function transactionReducer(
   state: TransactionState,
@@ -6,15 +6,14 @@ export function transactionReducer(
 ): TransactionState {
   switch (action.type) {
     case 'SET_TRANSACTIONS':
-      return { ...state, transactions: action.payload };
+      return { ...state, transactions: [...state.transactions, ...action.payload] };
     case 'SET_PARSED_DATA':
-      return { ...state, parsedData: action.payload };
+      return { ...state, parsedData: [...state.parsedData, ...action.payload] };
     case 'SET_BROKEN_DATA':
-      return { ...state, brokenData: action.payload };
+      return { ...state, brokenData: [...state.brokenData, ...action.payload] };
     case 'SET_PAGE':
       return { ...state, currentPage: action.payload };
     default:
       return state;
   }
 }
-
