+++ transaction-processor/src/app/page.tsx
@@ -1,30 +1,22 @@
 'use client'
 
 // TO DO
-// A list of "bad transactions" that someone needs to go look at (any transactions that you were unable to parse)
-// NOT SURE HOW TO QUALIFY "BAD TRANS". break data.
+// write readme
+// write tests
+// allow csvs and other file formats
 
-// Each file submitted should continue "in continuation of" previous submissions
-// validation for file type
-
+// FINAL QA
+// check file naming conventions align with next js standards
+// make sure numbers don't float
+// format all instances of USD
 // remove comments and console logs
 
-// write readme
-// conditionally render button text based on if file has been uploaded or not
-// clone to standalone repo
-// deploy to vercel
-// highlight button of selected section
-// favicon
+// NICE TO HAVE
 // paginate table OR at least lazy load data
-// check file naming conventions align with next js standards
-// <!-- allow csvs and other file formats -->
-// make sure numbers don't float
-// write tests
-
-// persistence during a "run" of the software is required
-// - for example, if you choose to build a nextjs or remix site, we expect that you, at minimum, use an in memory cache that maintains state as long as the process is running
 // fix overlap of collapsed table cells
-// format all instances of USD
+// favicon
+// highlight button of selected section
+// conditionally render button text based on if file has been uploaded or not
 
 import { TransactionProvider } from './context/context';
 import WidthContainer from './components/width-container/widthContainer';
